changed every var to decimal but you need to replace the string ',' with '.'

+++ Smart_Orders_Project/Smart_Orders_Project/Models/LineOfOrder.cs
@@ -10,10 +10,10 @@ namespace Smart_Orders_Project.Models
         public Product Product { get; set; }
         public Guid RFSalesOid { get; set; }
         public string ProductBarCode { get; set; }
-        public double Quantity { get; set; }
-        public double Width { get; set; }
-        public double Length { get; set; }
-        public double Height { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Width { get; set; }
+        public decimal Length { get; set; }
+        public decimal Height { get; set; }
         public double Sum { get; set; }
     }
 }
